Reorganize python tests
