+++ test/python/test_processes.py
@@ -1,0 +1,118 @@
+"""
+Test composite classes
+"""
+import pytest
+import reactions
+
+
+def test_reaction():
+
+    reac = reactions.reaction('pi+ -> mu+ nu_mu')
+
+    # check sizes
+    assert len(reac.reactants) == 1
+    assert len(reac.products) == 2
+
+    # check types
+    assert reactions.node_type(reac) == 'reaction'
+    for e in reac.reactants + reac.products:
+        assert reactions.node_type(e) == 'element'
+
+    # more complex reactions
+    reactions.reaction('p+ p~- -> {pi+ -> mu+ nu_mu} pi-')
+
+    # we can create empty reactions
+    reac = reactions.reaction(kind='pdg')
+
+    with pytest.raises(ValueError):
+        reactions.reaction(kind='nothing')
+
+    # we can append new elements
+    reac.reactants.append(reactions.pdg_element('pi+'))
+    reac.products.append(reactions.pdg_element('mu+'))
+    reac.products.append(reactions.pdg_element('nu_mu'))
+
+    # comparison operators
+    assert reac == reac
+
+    with pytest.raises(TypeError):
+        # different element kinds
+        reac != reactions.reaction('pi- -> mu- nu_mu~')
+
+    reac != reactions.reaction('pi- -> mu- nu_mu~', kind='pdg')
+
+    # order does not matter
+    assert reac == reactions.reaction('pi+ -> nu_mu~ mu-', kind='pdg')
+    reac = reactions.reaction('p+ p~- -> {pi+ -> mu+ nu_mu} pi-')
+    assert reac == reactions.reaction('p~- p+ -> pi- {pi+ -> nu_mu mu+}')
+
+
+def test_decay():
+
+    reac = reactions.decay('pi+ -> mu+ nu_mu')
+
+    # check products size
+    assert len(reac.products) == 2
+
+    # check types
+    assert reactions.node_type(reac) == 'decay'
+    for e in reac.products:
+        assert reactions.node_type(e) == 'element'
+
+    # more complex reactions
+    reactions.decay('p+ -> {pi+ -> mu+ nu_mu} pi-')
+
+    # we can create empty reactions
+    reac = reactions.decay(kind='pdg')
+
+    with pytest.raises(ValueError):
+        reactions.decay(kind='nothing')
+
+    # we can append new elements
+    reac.head = reactions.pdg_element('pi+')
+    reac.products.append(reactions.pdg_element('mu+'))
+    reac.products.append(reactions.pdg_element('nu_mu'))
+
+    # comparison operators
+    assert reac == reac
+
+    with pytest.raises(TypeError):
+        # different element kinds
+        reac != reactions.decay('pi- -> mu- nu_mu~')
+
+    assert reac != reactions.decay('pi- -> mu- nu_mu~', kind='pdg')
+
+    # order does not matter
+    assert reac == reactions.decay('pi+ -> nu_mu~ mu-', kind='pdg')
+    reac = reactions.decay('KS0 -> {pi+ -> mu+ nu_mu} pi-')
+    assert reac == reactions.decay('KS0 -> pi- {pi+ -> nu_mu mu+}')
+
+
+def test_syntax():
+
+    for proc in reactions.reaction, reactions.decay:
+
+        with pytest.raises(RuntimeError):
+            proc('A ->')
+
+        with pytest.raises(RuntimeError):
+            proc('-> B C')
+
+        with pytest.raises(RuntimeError):
+            proc('{A -> B C')
+
+        with pytest.raises(RuntimeError):
+            proc('A -> B C}')
+
+        with pytest.raises(RuntimeError):
+            proc('A -> -> B C')
+
+        with pytest.raises(RuntimeError):
+            proc('A -> B {C')
+
+    with pytest.raises(RuntimeError):
+        reactions.decay('A B -> C D')
+
+    reactions.reaction('A -> B {C D -> E} F')
+    with pytest.raises(RuntimeError):
+        reactions.decay('A -> B {C D -> E} F')
